backend/requirements.txt
streamlit>=1.24.0
openai>=1.0.0
numpy>=1.24.0
pypdf2
pandas
Flask
Flask-Cors
python-dotenv
